+++ app/src/main/java/com/example/catapult/data/ui/LBItemUiModel.kt
@@ -3,6 +3,6 @@ package com.example.catapult.data.ui
 data class LBItemUiModel (
     val ranking: Int = -1,
     val nickname: String = "",
-    val result: Float = 0f,
+    val result: Double = 0.0,
     val totalGamesPlayed: Int = 0,
 )
